add fitler in avenue , cities

+++ app/Http/Controllers/Country/CountryController.php
@@ -67,9 +67,10 @@ class CountryController extends Controller
         if (!$model) {
             return responseJson(404, __('message.data not found'));
         }
-        if ($model->governorates()->count() > 0) {
-            return responseJson(400, __('message.country has governorates'));
+        if ($model->governorates()->count() > 0 || $model->avenues()->count() > 0) {
+            return responseJson(400, __('message.country has sub data'));
         }
+
         $this->modelInterface->delete($id);
 
         return responseJson(200, 'success');
